move constraints to EUDEMO
